backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.14.0
Pillow==10.4.0
python-decouple==3.8
whitenoise==6.8.2
python-dateutil==2.9.0
requests==2.32.3
gunicorn==23.0.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
# Dependências pesadas removidas para otimizar memória:
# WeasyPrint - biblioteca pesada para PDF
# PyMuPDF - biblioteca pesada para manipulação de PDF
# django-extensions - não essencial para produção
# python-docx - biblioteca pesada para Word
# openpyxl - biblioteca pesada para Excel
django-storages==1.14.4
boto3==1.35.84
